+++ spring_notes/src/main/java/spring_notes/entities/Note.java
@@ -1,4 +1,4 @@
-package spring_notes.note.entity;
+package spring_notes.entities;
 
 import javax.persistence.*;
 import java.sql.Timestamp;
