+++ spring_notes/src/main/java/spring_notes/repositories/NoteRepository.java
@@ -1,7 +1,7 @@
-package spring_notes.note.repository;
+package spring_notes.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import spring_notes.note.entity.Note;
+import spring_notes.entities.Note;
 
 public interface NoteRepository extends JpaRepository<Note,Long> {
 }
