+++ spring_notes/src/main/java/spring_notes/repositories/PersonRepository.java
@@ -1,7 +1,7 @@
-package spring_notes.person.repository;
+package spring_notes.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import spring_notes.person.entity.Person;
+import spring_notes.entities.Person;
 
 public interface PersonRepository extends JpaRepository<Person,Long> {
 }
