--- spring_notes/src/main/resources/application-postgres.yml
@@ -1,19 +1,0 @@
-spring:
-  profiles:
-    active: prod
-  datasource:
-    url: jdbc:postgresql://localhost:5432/notes
-    driver-class-name: org.postgresql.Driver
-    username: postgres
-    password: postgres
-  jpa:
-    hibernate:
-      ddl-auto: create-drop
-    show-sql: true
-    properties:
-      hibernate:
-        jdbc:
-          lob:
-            non_contextual_creation: true
-            database-platform: org.hibernate.dialect.PostgreSQL9Dialect
-        format_sql: true
