+++ spring_notes/src/main/resources/application.yml
@@ -1,21 +1,19 @@
 spring:
   profiles:
-    active: dev
+    active: prod
   datasource:
-    url: jdbc:h2:mem:memDb;DB_CLOSE_DELAY=-1
-    driver-class-name: org.h2.Driver
-    username: sa
-    password: sa
-  sql:
-    init:
-      schema-locations: classpath:schema.sql
-  h2:
-    console:
-      enabled: true
+    url: jdbc:postgresql://localhost:5432/notes
+    driver-class-name: org.postgresql.Driver
+    username: postgres
+    password: postgres
   jpa:
     hibernate:
-      ddl-auto: validate
+      ddl-auto: update
     show-sql: true
     properties:
       hibernate:
-        dialect: org.hibernate.dialect.H2Dialect
+        jdbc:
+          lob:
+            non_contextual_creation: true
+            database-platform: org.hibernate.dialect.PostgreSQL9Dialect
+        format_sql: true
